+++ NEWS.md
@@ -1,5 +1,6 @@
 Version 0.6.8
  - enhancement: add link to cppreference in the help menu
+ - fix: add mutex lock to prevent editor crash in rare conditions
 
 Version 0.6.7
  - fix: messages send to the gdb process's standard error are not received